+++ app/widgets/Visio/visio.js
@@ -1,209 +1,23 @@
-/*function logError(error) {
+function logError(error) {
     console.log(error.name + ': ' + error.message);
-    console.log(error);
+    console.error(error);
 }
 
 var Visio = {
-    pc: null,
+    from: null,
+    id: null,
     localVideo: null,
     remoteVideo: null,
-    constraints: null,
-    audioContext: null,
-    max_level_L: 0,
-    old_level_L: 0,
-    from: null,
+
+    calling: false,
 
     videoSelect: undefined,
     switchCamera: undefined,
 
-    setFrom: from => Visio.from = from,
-
-    gotStream: function() {
-        Visio.constraints = window.constraints = {
-            audio: true,
-            video: {
-                deviceId: Visio.videoSelect.value,
-                facingMode: 'user',
-                width: { ideal: 4096 },
-                height: { ideal: 4096 }
-            }
-        };
-
-        Visio.switchCamera.classList.add('disabled');
-        Visio.localVideo.srcObject = null;
-
-        // Useful on Android where you can't have both camera enabled at the same time
-        var videoTrack = Visio.pc.getSenders().find(rtc => rtc.track && rtc.track.kind == 'video');
-        if (videoTrack) videoTrack.track.stop();
-
-        if(navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
-            navigator.mediaDevices.getUserMedia(Visio.constraints)
-            .then(function(stream) {
-                Visio.switchCamera.classList.remove('disabled');
-
-                Visio.localVideo.srcObject = stream;
-                Visio.handleAudio();
-
-                // For the first time we attach all the tracks
-                if (Visio.pc.getSenders().length == 0) {
-                    stream.getTracks().forEach(track => Visio.pc.addTrack(track, stream));
-                }
-
-                // Switch camera
-                let videoTrack = stream.getVideoTracks()[0];
-                var sender = Visio.pc.getSenders().find(s => s.track && s.track.kind == videoTrack.kind);
-
-                if (sender) {
-                    sender.replaceTrack(videoTrack);
-                }
-
-                Visio.toggleMainButton();
-
-                // If we received an offer, we need to answer
-                if (Visio.pc.remoteDescription.type == 'offer') {
-                    Visio.pc.createAnswer(Visio.localDescCreated, logError);
-                }
-            });
-        }
-    },
-
-    gotDevices: function(deviceInfos) {
-        Visio.videoSelect.innerText = '';
-
-        const ids = [];
-
-        for (let i = 0; i !== deviceInfos.length; ++i) {
-            const deviceInfo = deviceInfos[i];
-
-            if (deviceInfo.kind === 'videoinput' && !ids.includes(deviceInfo.deviceId)) {
-                const option = document.createElement('option');
-                option.value = deviceInfo.deviceId;
-                option.text = deviceInfo.label;
-                Visio.videoSelect.add(option);
-                ids.push(deviceInfo.deviceId);
-            }
-        }
-
-        if (ids.length >= 2) {
-            document.querySelector("#visio #switch_camera").classList.add('enabled');
-        }
-
-        Visio.videoSelect.addEventListener('change', e => Visio.gotStream());
-        Visio.gotStream();
-    },
-
-    handleAudio: function(stream) {
-        Visio.audioContext = new AudioContext();
-
-        var microphone = Visio.audioContext.createMediaStreamSource(Visio.localVideo.srcObject);
-        var javascriptNode = Visio.audioContext.createScriptProcessor(2048, 1, 1);
-
-        var cnvs = document.querySelector('#visio .level');
-        var cnvs_cntxt = cnvs.getContext('2d');
-
-        microphone.connect(javascriptNode);
-        javascriptNode.connect(Visio.audioContext.destination);
-        javascriptNode.onaudioprocess = function(event) {
-            var inpt_L = event.inputBuffer.getChannelData(0);
-            var instant_L = 0.0;
-
-            var sum_L = 0.0;
-
-            for(var i = 0; i < inpt_L.length; ++i) {
-                sum_L += inpt_L[i] * inpt_L[i];
-            }
-
-            instant_L = Math.sqrt(sum_L / inpt_L.length);
-            Visio.max_level_L = Math.max(Visio.max_level_L, instant_L);
-            instant_L = Math.max( instant_L, Visio.old_level_L -0.008 );
-            Visio.old_level_L = instant_L;
-
-            cnvs_cntxt.clearRect(0, 0, cnvs.width, cnvs.height);
-            cnvs_cntxt.fillStyle = 'white';
-            cnvs_cntxt.fillRect(0, 0,(cnvs.width)*(instant_L/Visio.max_level_L),(cnvs.height)); // x,y,w,h
-        }
-    },
-
-    onSDP: function(sdp, type) {
-        console.log('SDP');
-        console.log(type);
-
-        Visio.pc.setRemoteDescription(
-            new RTCSessionDescription({'sdp': sdp + "\n", 'type': type}),
-            () => {
-                Visio_ajaxGetCandidates();
-            },
-            (error) => {
-                Visio.goodbye('incompatible-parameters');
-                logError(error);
-            }
-        );
-    },
-
-    localDescCreated: function(desc) {
-        Visio.pc.setLocalDescription(desc, () => {
-            Visio_ajaxAccept(Visio.pc.localDescription, Visio.from);
-            Visio.toggleMainButton();
-        }, logError);
-    },
-
-    onCandidates: function(candidates) {
-        console.log('Canditates');
-        candidates.forEach(candidate => Visio.onCandidate(candidate[0], candidate[1], candidate[2]));
-    },
-
-    onCandidate: function(candidate, mid, mlineindex) {
-        console.log('candidate');
-        console.log(candidate);
-
-        Visio_ajaxGetCandidates();
-
-        if (mid == '') mlineindex = 1;
-
-        if (Visio.pc.remoteDescription == null || candidate.candidate == '') return;
-
-        candidate = new RTCIceCandidate({
-            'candidate': candidate,
-            'sdpMid': mid,
-            'sdpMLineIndex' : mlineindex
-        });
-
-        Visio.pc.addIceCandidate(candidate, e => {}, logError);
-    },
-
-    onTerminate: () => {
-        let localStream = Visio.localVideo.srcObject;
-
-        if (localStream) {
-            localStream.getTracks().forEach(track => track.stop());
-        }
-
-        let remoteStream = Visio.remoteVideo.srcObject;
-
-        if (remoteStream) {
-            remoteStream.getTracks().forEach(track => track.stop());
-        }
-
-        Visio.localVideo.srcObject = null;
-        Visio.remoteVideo.srcObject = null;
-        Visio.localVideo.classList.add('hide');
-
-        if (Visio.pc) Visio.pc.close();
-
-        document.querySelector('p.state').innerText = Visio.states.ended;
-        button = document.querySelector('#main');
-
-        button.className = 'button action color red';
-        button.querySelector('i').className = 'material-icons';
-        button.querySelector('i').innerText = 'close';
-
-        button.onclick = () => window.close();
-    },
-
-    init: (sdp, type) => {
-        Visio.toggleMainButton();
-
-        Visio.setFrom(MovimUtils.urlParts().params.join('/'));
+    init: function(id) {
+        Visio.from = MovimUtils.urlParts().params[0] + '/' + MovimUtils.urlParts().params[1];
+        Visio.localVideo = document.getElementById('video');
+        Visio.remoteVideo = document.getElementById('remote_video');
 
         const servers = ['stun:stun01.sipphone.com',
             'stun:stun.ekiga.net',
@@ -236,23 +50,18 @@ var Visio = {
 
         Visio.pc = new RTCPeerConnection(configuration);
 
-        Visio.pc.ontrack = (event) => {
-            const stream = event.streams[0];
-            if (!Visio.remoteVideo.srcObject || Visio.remoteVideo.srcObject.id !== stream.id) {
-                Visio.remoteVideo.srcObject = stream;
-                Visio.remoteVideo.classList.add('enabled');
-            }
+        Visio.pc.ontrack = event => {
+            Visio.remoteVideo.srcObject = event.streams[0];
         };
 
-        Visio.pc.onicecandidate = (evt) => {
-            Visio.toggleMainButton();
-
+        Visio.pc.onicecandidate = evt => {
             if (evt.candidate) {
+                console.log(evt.candidate);
                 Visio_ajaxCandidate(evt.candidate, Visio.from);
             }
         };
 
-        Visio.pc.oniceconnectionstatechange = () => Visio.toggleMainButton();
+        Visio.pc.oniceconnectionstatechange = () => VisioUtils.toggleMainButton();
 
         Visio.pc.onicegatheringstatechange = function (e) {
             // When we didn't receive the WebRTC termination before Jingle
@@ -260,167 +69,151 @@ var Visio = {
                 Visio.onTerminate();
             }
 
-            Visio.toggleMainButton();
+            VisioUtils.toggleMainButton();
         };
 
-        Visio.toggleMainButton();
+        VisioUtils.toggleMainButton();
 
-        if (sdp && type) {
-            Visio.onSDP(sdp, type);
-        }
-
-        // Switch camera
-        Visio.videoSelect = document.querySelector('#visio select#visio_source');
-        navigator.mediaDevices.enumerateDevices().then(devices => Visio.gotDevices(devices));
-
-        Visio.switchCamera = document.querySelector("#visio #switch_camera");
-        Visio.switchCamera.onclick = () => {
-            Visio.videoSelect.selectedIndex++;
-
-            // No empty selection
-            if (Visio.videoSelect.selectedIndex == -1) {
-                Visio.videoSelect.selectedIndex++;
-            }
-
-            Visio.gotStream();
-        };
+        VisioUtils.switchCameraSetup();
     },
 
-    hello: function() {
-        Visio.pc.createOffer((desc) => {
-            Visio.pc.setLocalDescription(
-                desc,
-                () => Visio_ajaxInitiate(Visio.pc.localDescription, Visio.from),
-                logError
-            );
+    gotStream: function() {
+        // On Android where you can't have both camera enabled at the same time
+        var videoTrack = Visio.pc.getSenders().find(rtc => rtc.track && rtc.track.kind == 'video');
+        if (videoTrack) videoTrack.track.stop();
+
+        Visio.switchCamera.classList.add('disabled');
+
+        navigator.mediaDevices.getUserMedia({
+            audio: true,
+            video: {
+                deviceId: Visio.videoSelect.value,
+                facingMode: 'user',
+                width: { ideal: 4096 },
+                height: { ideal: 4096 }
+            }
+        }).then(stream => {
+            Visio.switchCamera.classList.remove('disabled');
+
+            Visio.localVideo.srcObject = stream;
+            VisioUtils.handleAudio();
+
+            // Switch camera
+            let videoTrack = stream.getVideoTracks()[0];
+            var sender = Visio.pc.getSenders().find(s => s.track && s.track.kind == videoTrack.kind);
+
+            if (sender) {
+                sender.replaceTrack(videoTrack);
+            }
+
+            VisioUtils.toggleMainButton();
+
+            // For the first time we attach all the tracks and we launch the call
+            if (Visio.pc.getSenders().length == 0) {
+                stream.getTracks().forEach(track => Visio.pc.addTrack(track, stream));
+
+                if (Visio.id) {
+                    Visio_ajaxAccept(Visio.from, Visio.id);
+                } else {
+                    // TODO launch when button pressed
+                    Visio.id = Math.random().toString(36).substr(2, 9);
+                    Visio.calling = true;
+                    VisioUtils.toggleMainButton();
+                    Visio_ajaxPropose(Visio.from, Visio.id);
+                }
+            }
         }, logError);
     },
 
-    goodbye: (reason) => {
-        Visio.onTerminate();
-        Visio_ajaxTerminate(Visio.from, reason);
+    onCandidate: function(candidate, mid, mlineindex) {
+        Visio.pc.addIceCandidate(new RTCIceCandidate({
+            'candidate': candidate.substr(2),
+            'sdpMid': mid,
+            'sdpMLineIndex' : mlineindex
+        }), e => {}, logError);
     },
 
-    toggleMainButton: function() {
-        button = document.getElementById('main');
-        state = document.querySelector('p.state');
-
-        i = button.querySelector('i');
-
-        button.classList.remove('red', 'green', 'gray', 'orange', 'ring', 'blue');
-        button.classList.add('disabled');
-
-        if (Visio.pc) {
-            let length = Visio.pc.getSenders().length;
-
-            if (Visio.pc.iceConnectionState != 'closed'
-            && length > 0) {
-                button.classList.remove('disabled');
-            }
-
-            button.onclick = function() {};
-
-            if (length == 0) {
-                button.classList.add('gray');
-                i.innerText = 'more_horiz';
-            } else if (Visio.pc.iceConnectionState == 'new') {
-                if (Visio.pc.iceGatheringState == 'gathering'
-                || Visio.pc.iceGatheringState == 'complete') {
-                    button.classList.add('orange');
-                    i.className = 'material-icons ring';
-                    i.innerText = 'call';
-                    state.innerText = Visio.states.ringing;
-
-                    button.onclick = function() { Visio.goodbye(); };
-                } else {
-                    button.classList.add('green');
-                    i.innerText = 'call';
-
-                    button.onclick = function() { Visio.hello(); };
-                }
-            } else if (Visio.pc.iceConnectionState == 'checking') {
-                button.classList.add('blue');
-                i.className = 'material-icons disabled';
-                i.innerText = 'more_horiz';
-                state.innerText = Visio.states.connecting;
-            } else if (Visio.pc.iceConnectionState == 'closed') {
-                button.classList.add('gray');
-                button.classList.remove('disabled');
-                i.innerText = 'call_end';
-
-                button.onclick = function() { Visio.goodbye(); };
-            } else if (Visio.pc.iceConnectionState == 'connected'
-                   || Visio.pc.iceConnectionState == 'complete'
-                   || Visio.pc.iceConnectionState == 'failed') {
-                button.classList.add('red');
-                i.className = 'material-icons';
-                i.innerText = 'call_end';
-
-                if (Visio.pc.iceConnectionState == 'failed') {
-                    state.innerText = Visio.states.failed;
-                } else {
-                    state.innerText = Visio.states.in_call;
-                }
-
-                button.onclick = () => Visio.goodbye();
-            }
-        }
-    },
-
-    toggleFullScreen: function() {
-        var button = document.querySelector('#toggle_fullscreen i');
-
-        if (!document.fullscreenElement) {
-            if (document.body.requestFullscreen) {
-                document.body.requestFullscreen();
-            }
-
-            button.innerText = 'fullscreen_exit';
+    onAccept: function(from, id) {
+        if (Visio.from == from && Visio.id == id) {
+            Visio.pc.createOffer().then(function(offer) {
+                Visio.calling = false;
+                VisioUtils.toggleMainButton();
+                return Visio.pc.setLocalDescription(offer);
+            })
+            .then(function() {
+                Visio_ajaxSessionInitiate(Visio.pc.localDescription, Visio.from, Visio.id);
+            });
         } else {
-            if (document.exitFullscreen) {
-                document.exitFullscreen();
-            }
-
-            button.innerText = 'fullscreen';
+            console.error('Wrong call')
         }
     },
 
-    toggleAudio: function() {
-        var button = document.querySelector('#toggle_audio i');
-        var rtc = Visio.pc.getSenders().find(rtc => rtc.track && rtc.track.kind == 'audio');
-
-        if (rtc && rtc.track.enabled == 1) {
-            rtc.track.enabled = 0;
-            button.innerText = 'mic_off';
-        } else if (rtc) {
-            rtc.track.enabled = 1;
-            button.innerText = 'mic';
-        }
+    onInitiateSDP: function(sdp) {
+        Visio.pc.setRemoteDescription(new RTCSessionDescription({'sdp': sdp + "\n", 'type': 'offer'}), () => {
+            Visio.pc.createAnswer().then(function(answer) {
+                return Visio.pc.setLocalDescription(answer);
+            }).then(function() {
+                Visio_ajaxSessionAccept(Visio.pc.localDescription, Visio.from, Visio.id);
+            }).catch(logError);
+        }, logError);
     },
 
-    toggleVideo: function() {
-        var button = document.querySelector('#toggle_video i');
-        var rtc = Visio.pc.getSenders().find(rtc => rtc.track && rtc.track.kind == 'video');
-
-        if (rtc) {
-            if (rtc.track.enabled == 1) {
-                rtc.track.enabled = 0;
-                button.innerText = 'videocam_off';
-            } else {
-                rtc.track.enabled = 1;
-                button.innerText = 'videocam';
+    onAcceptSDP: function(sdp) {
+        Visio.pc.setRemoteDescription(
+            new RTCSessionDescription({'sdp': sdp + "\n", 'type': 'answer'}), () => {},
+            (error) => {
+                Visio.goodbye('incompatible-parameters');
+                logError(error)
             }
+        );
+    },
+
+    onTerminate: (reason) => {
+        console.log(reason);
+        let localStream = Visio.localVideo.srcObject;
+
+        if (localStream) {
+            localStream.getTracks().forEach(track => track.stop());
         }
+
+        let remoteStream = Visio.remoteVideo.srcObject;
+
+        if (remoteStream) {
+            remoteStream.getTracks().forEach(track => track.stop());
+        }
+
+        Visio.localVideo.srcObject = null;
+        Visio.remoteVideo.srcObject = null;
+        Visio.localVideo.classList.add('hide');
+
+        if (Visio.pc) Visio.pc.close();
+
+        document.querySelector('p.state').innerText = reason == 'decline'
+            ? Visio.states.declined
+            : Visio.states.ended;
+        button = document.querySelector('#main');
+
+        button.className = 'button action color red';
+        button.querySelector('i').className = 'material-icons';
+        button.querySelector('i').innerText = 'close';
+
+        button.onclick = () => window.close();
+    },
+
+    goodbye: (reason) => {
+        Visio.onTerminate(reason);
+        Visio_ajaxTerminate(Visio.from, reason);
     },
 }
 
 MovimWebsocket.attach(() => {
-    Visio.localVideo = document.getElementById('video');
-    Visio.remoteVideo = document.getElementById('remote_video');
-    Visio_ajaxAskInit();
+    if (MovimUtils.urlParts().params[2] !== undefined) {
+        Visio.id = MovimUtils.urlParts().params[2];
+    }
+
+    Visio.init();
 });
 
 window.onbeforeunload = () => {
     Visio.goodbye();
 }
-*/