+++ lib/moxl/src/Moxl/Stanza/Jingle.php
@@ -4,6 +4,38 @@ namespace Moxl\Stanza;
 
 class Jingle
 {
+    public static function sessionPropose($to, $id)
+    {
+        $dom = new \DOMDocument('1.0', 'UTF-8');
+        $message = $dom->createElementNS('jabber:client', 'message');
+        $message->setAttribute('to', $to);
+        $dom->appendChild($message);
+
+        $propose = $dom->createElementNS('urn:xmpp:jingle-message:0', 'propose');
+        $propose->setAttribute('id', $id);
+        $message->appendChild($propose);
+
+        $description = $dom->createElementNS('urn:xmpp:jingle:apps:rtp:1', 'description');
+        $description->setAttribute('media', 'video');
+        $propose->appendChild($description);
+
+        \Moxl\API::request($dom->saveXML($dom->documentElement));
+    }
+
+    public static function sessionAccept($to, $id)
+    {
+        $dom = new \DOMDocument('1.0', 'UTF-8');
+        $message = $dom->createElementNS('jabber:client', 'message');
+        $message->setAttribute('to', $to);
+        $dom->appendChild($message);
+
+        $propose = $dom->createElementNS('urn:xmpp:jingle-message:0', 'accept');
+        $propose->setAttribute('id', $id);
+        $message->appendChild($propose);
+
+        \Moxl\API::request($dom->saveXML($dom->documentElement));
+    }
+
     public static function sessionInitiate($to, $offer)
     {
         $xml = \Moxl\API::iqWrapper($offer, $to, 'set');