+++ lib/moxl/src/Moxl/Xec/Action/Jingle/SessionPropose.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Moxl\Xec\Action\Jingle;
+
+use Moxl\Xec\Action;
+use Moxl\Stanza\Jingle;
+
+class SessionPropose extends Action
+{
+    protected $_to;
+    protected $_id;
+
+    public function request()
+    {
+        $this->store();
+        Jingle::sessionPropose($this->_to, $this->_id);
+    }
+}