+++ lib/moxl/src/Moxl/Xec/Payload/JingleAccept.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Moxl\Xec\Payload;
+
+class JingleAccept extends Payload
+{
+    public function handle($stanza, $parent = false)
+    {
+        $this->pack([
+            'from' => (string)$parent->attributes()->from,
+            'id' => (string)$stanza->attributes()->id
+        ]);
+        $this->deliver();
+    }
+}